+++ src/js/menu/settings/index.js
@@ -2,14 +2,14 @@ import { h } from "preact";
 import { MenuSection } from "@/components/menuItems";
 import SpacebarMute from '@/menu/settings/spacebar-mute';
 import WarnOnNavigation from '@/menu/settings/warn-on-navigation';
-import filterAddToPlaylists from '@/menu/settings/filter-add-to-playlists';
+import FilterAddToPlaylists from '@/menu/settings/filter-add-to-playlists';
 
 const SettingsSection = () => {
   return (
     <MenuSection id="dubplus-settings" title="Settings" settingsKey="settings">
       <SpacebarMute />
       <WarnOnNavigation />
-      <filterAddToPlaylists />
+      <FilterAddToPlaylists />
     </MenuSection>
   );
 };